Build a notebook available locally but not in *.tar.gz
The script will now check if the provided filename is a locally
available directory. This is useful if you sync your notebooks
with ownCloud.
The following file structure is expected:

- $filename/{*.rm,*.json}
- $filename.content
- $filename.metadata
- $filename.pagedata

To provide the locally required template files (available on
the RM at /usr/share/remarkable/templates) use the -t parameter.